sdk/python/requirements.txt
# OpenADP Python SDK Dependencies
# 
# Core dependencies for cryptographic operations and network communication

# Cryptographic operations
cryptography>=41.0.0
noiseprotocol>=0.3.1

# HTTP client for JSON-RPC communication
requests>=2.28.0

# Development and testing dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Type checking (optional)
mypy>=1.0.0
types-requests>=2.28.0

# Code formatting (optional)
black>=23.0.0
isort>=5.12.0

# Documentation (optional)
sphinx>=6.0.0
sphinx-rtd-theme>=1.3.0 